backend/requirements.txt
fastapi
uvicorn
pika
python-multipart
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic[email]
pydantic-settings
aiofiles
sqlalchemy
alembic
clarifai
python-magic
minio
celery
redis
pillow